Docker container to serve cloudefficiency out of ecs.
