+++ .travis.yml
@@ -1,0 +1,18 @@
+matrix:
+  include:
+    - language: python
+      python: 3.6
+      before_script:
+        - cd report
+
+    - language: python
+      python: 3.6
+      before_script:
+        - cd server
+
+    - language: node_js
+      node_js:
+        - "iojs"
+        - "7"
+      before_script:
+        - cd frontend
