--- server/authorizer/requirements.txt
@@ -1,4 +1,0 @@
-requests-oauthlib==1.0.0
-PyJWT==1.6.4
-boto3==1.7.20
-botocore==1.10.38
