+++ server/cloudformation/buckets.yaml
@@ -1,0 +1,32 @@
+AWSTemplateFormatVersion: 2010-09-09
+Description: >
+  Buckets for cpe keymaster/ECS static site serving.
+Parameters:
+  StaticBucketName:
+    Type: String
+    Description: The name to use for the static assets bucket.
+  LoggingBucketName:
+    Type: String
+    Description: The name to use for the elb, cloudfront etc. logging bucket.
+Resources:
+  StaticBucket:
+    Type: AWS::S3::Bucket
+    Properties:
+      BucketName: !Ref StaticBucketName
+  LoggingBucket:
+    Type: AWS::S3::Bucket
+    Properties:
+      BucketName: !Ref LoggingBucketName
+Outputs:
+  StaticBucket:
+    Description: Static asset bucket for cloudefficiency.
+    Value: !Ref StaticBucket
+    Export:
+      Name:
+        Fn::Sub: "${AWS::StackName}-StaticBucket"
+  LoggingBucket:
+    Description: logging bucket for ELB, cloudfront, etc.
+    Value: !Ref LoggingBucket
+    Export:
+      Name:
+        Fn::Sub: "${AWS::StackName}-LoggingBucket"
