+++ server/cloudformation/ecr_repo.yaml
@@ -1,0 +1,10 @@
+AWSTemplateFormatVersion: "2010-09-09"
+Description: IAM resources for CPE keymaster serving.
+Parameters:
+  RepoName:
+    Type: String
+Resources:
+  Repo:
+    Type: AWS::ECR::Repository
+    Properties:
+      RepositoryName: !Ref RepoName
