+++ server/cloudformation/redirect_index.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="utf-8">
+  <meta name="viewport"
+     content="width=device-width, initial-scale=1, user-scalable=yes">
+ 
+  <title>Hello World!</title>
+</head>
+<body>
+Hello World!
+ 
+</body>
+</html>
