+++ server/requirements.txt
@@ -1,3 +1,4 @@
+Flask==1.0.2
 requests-oauthlib==1.0.0
 PyJWT==1.6.4
 boto3==1.7.20
