--- server/stagevariables.json.example
@@ -1,11 +1,0 @@
-'client_id': 'something',
-'client_secret': 'something2',
-'authorization_endpoint': 'https://example.com/authorize',
-'token_endpoint': 'https://example.com/token',
-'userinfo_endpoint': 'https://example.com/userinfo',
-'redirect_uri': 'https://<yourwebsite>/callback',
-'base_url': 'https://<yourwebsite>',
-'jwt_secrets': 'secret,secret2'
-
-
-
